+++ src/components/product-menu/__tests__/product-menu-test-cases.js
@@ -1,50 +1,14 @@
 import React from 'react';
 import ProductMenu from '../product-menu';
-import { ProductMenuItems } from '../../../data/product-menu-items';
-import ProductMenuDropdown from '../../product-menu-dropdown/product-menu-dropdown';
 
 const testCases = {};
-const locationTestItems = [
-  {
-    productCategory: 'Foo',
-    icon: 'map',
-    products: [
-      {
-        url: '/mapbox-gl-js/',
-        name: 'Foo1'
-      },
-      {
-        url: '/ios-sdk/maps/',
-        name: 'Foo2'
-      }
-    ]
-  },
-  {
-    productCategory: 'Bar',
-    icon: 'compass',
-    products: [
-      {
-        url: '/ProductMenu',
-        name: 'ProductMenu'
-      }
-    ]
-  }
-];
-
-function renderMenu(items) {
-  return <ProductMenuDropdown categories={items} />;
-}
-
-function renderBoring() {
-  return <img src={'https://i.giphy.com/media/A9lgUYVqLeRb2/giphy.webp'} />;
-}
 
 testCases.mapboxMenu = {
   component: ProductMenu,
   description: 'Mapbox product menu',
   props: {
     productName: 'Mapbox product names',
-    children: renderMenu(ProductMenuItems)
+    homePage: '/api-documentation/'
   }
 };
 
@@ -53,7 +17,7 @@ testCases.locationMenu = {
   description: "Fake items, but one matches this test case's location",
   props: {
     productName: 'Items in here',
-    children: renderMenu(locationTestItems)
+    homePage: '/api-documentation/'
   }
 };
 
@@ -62,7 +26,7 @@ testCases.arbitraryElement = {
   description: 'Product menu with something arbitrary in it',
   props: {
     productName: 'Pizza pie',
-    children: renderBoring()
+    homePage: '/api-documentation/'
   }
 };
 
@@ -70,9 +34,11 @@ testCases.customStyle = {
   description: 'Custom trigger style - display only',
   element: (
     <div className="bg-gray-dark py6">
-      <ProductMenu productName="Light on dark" lightText={true}>
-        {renderMenu(locationTestItems)}
-      </ProductMenu>
+      <ProductMenu
+        productName="Light on dark"
+        lightText={true}
+        homePage="/api-documentation/"
+      />
     </div>
   )
 };