+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Master
+
+- Removes ProductMenuDropdown and ProductMenuItems and refactors ProductMenu as a link. ([#73](https://github.com/mapbox/dr-ui/pull/73))
+
 ## 0.0.13
 
 - Add scrollspy feature to `NavigationAccordion` and add plugin to create sections in markdown files based on h2 and h3 headings. ([#66](https://github.com/mapbox/dr-ui/pull/66))